Extract the `WebViewServer`
